Refactor: indexPage, profilePage, listPage, errorPage를 한 클래스로 관리
Controller의 if문을 줄이기 위해 dynamicPage를 생성하는 하나의 메소드 추가
uri 매개변수에 따라 다른 메소드 호출
