+++ src/main/java/controller/Controller.java
@@ -6,10 +6,7 @@ import http.HttpRequest;
 import http.HttpResponse;
 import http.HttpStatus;
 import http.MIME;
-import view.IndexPage;
-import view.ListPage;
 import view.Page;
-import view.ProfilePage;
 
 import static exception.ExceptionList.INVALID_URI;
 import static http.Extension.HTML;
@@ -20,9 +17,6 @@ import static utils.FileIOUtils.*;
 
 public abstract class Controller {
     private final Page page = new Page();
-    private final ProfilePage profilePage = new ProfilePage();
-    private final IndexPage indexPage = new IndexPage();
-    private final ListPage listPage = new ListPage();
 
     public HttpResponse.ResponseBuilder loadFileByRequest(HttpRequest httpRequest) {
         try {
@@ -38,16 +32,8 @@ public abstract class Controller {
             if (getExtension().stream().noneMatch(entry -> entry.getKey().equals(extension))) {
                 throw new BadRequestException(INVALID_URI);
             }
-            if (extension.equals(HTML)) {
-                if (uri.endsWith(INDEX))
-                    return loadFileFromString(HttpStatus.OK, indexPage.getIndexPage(httpRequest), INDEX)
-                            .setContentType(MIME.getMIME().get(HTML));
-                if (uri.endsWith(PROFILE))
-                    return loadFileFromString(HttpStatus.OK, profilePage.getProfilePage(httpRequest), PROFILE)
-                            .setContentType(MIME.getMIME().get(HTML));
-                if (uri.endsWith(LIST))
-                    return loadFileFromString(HttpStatus.OK, listPage.getListPage(httpRequest), LIST)
-                            .setContentType(MIME.getMIME().get(HTML));
+            if (uri.endsWith(INDEX) || uri.endsWith(PROFILE) || uri.endsWith(LIST)) {
+                return loadFileFromString(HttpStatus.OK, page.getDynamicPage(httpRequest, uri), uri);
             }
             return loadFromPath(HttpStatus.OK, uri)
                     .setContentType(MIME.getMIME().get(extension));
@@ -55,8 +41,7 @@ public abstract class Controller {
             return loadFromPath(e.getHttpStatus(), e.getFilePath())
                     .setContentType(MIME.getMIME().get(HTML));
         } catch (BadRequestException e) {
-            String errorPage = page.getErrorPage(e.getMessage());
-            return loadFileFromString(HttpStatus.NOT_FOUND, errorPage, ERROR)
+            return loadFileFromString(HttpStatus.NOT_FOUND, page.getErrorPage(e.getMessage()), ERROR)
                     .setContentType(MIME.getMIME().get(HTML));
         } catch (Exception e) {
             return null;
